Add an explicit step to take/review the photo
Also show the photo and video together on the review page.
